Add SuspenseWithPerf at lower levels when not using redux
