+++ examples/react-firebase-redux/src/routes/Account/components/AccountEditor/AccountEditor.styles.js
@@ -1,0 +1,13 @@
+export default theme => ({
+  avatarCurrent: {
+    width: '100%',
+    maxWidth: '13rem',
+    marginTop: '3rem',
+    height: 'auto',
+    cursor: 'pointer'
+  },
+  gridItem: {
+    textAlign: 'center',
+    marginTop: theme.spacing(5)
+  }
+})
