+++ examples/react-firebase-redux/src/routes/Account/components/AccountPage/AccountPage.styles.js
@@ -4,30 +4,13 @@ export default theme => ({
     paddingBottom: theme.spacing(3),
     overflowY: 'scroll'
   },
-  pane: {
-    ...theme.flexColumnCenter,
-    justifyContent: 'space-around',
-    padding: theme.spacing(6)
-  },
-  avatar: {
-    maxWidth: '13rem',
-    marginTop: '3rem'
-  },
-  avatarCurrent: {
-    width: '100%',
-    maxWidth: '13rem',
-    marginTop: '3rem',
-    height: 'auto',
-    cursor: 'pointer'
-  },
   gridItem: {
     textAlign: 'center',
     marginTop: theme.spacing(5)
   },
-  meta: {
+  pane: {
     ...theme.flexColumnCenter,
-    flexBasis: '60%',
-    marginBottom: '3rem',
-    marginTop: '2rem'
+    justifyContent: 'space-around',
+    padding: theme.spacing(6)
   }
 })
