+++ examples/react-firebase-redux/src/routes/Projects/routes/Project/components/ProjectData/index.js
@@ -1,0 +1,3 @@
+import ProjectData from './ProjectData'
+
+export default ProjectData
