+++ examples/react-firebase/src/routes/Projects/routes/Project/components/ProjectData/ProjectData.styles.js
@@ -1,0 +1,1 @@
+export default theme => ({})
