+++ examples/react-firebase/src/routes/Projects/routes/Project/components/ProjectPage/ProjectPage.styles.js
@@ -2,9 +2,7 @@ export default theme => ({
   root: {
     padding: theme.spacing(2)
   },
-  progress: {
-    ...theme.flexRowCenter,
-    alignItems: 'center',
-    paddingTop: theme.spacing(8)
+  card: {
+    marginBottom: theme.spacing(2)
   }
 })
