+++ examples/redux-firestore/src/routes/Account/components/AccountEditor/index.js
@@ -1,0 +1,3 @@
+import AccountEditor from './AccountEditor'
+
+export default AccountEditor
