+++ generators/app/templates/src/layouts/CoreLayout/CoreLayout.js
@@ -1,10 +1,10 @@
-import React<% if (!includeRedux) { %>, { Suspense }<% } %> from 'react'
+import React from 'react'
 import PropTypes from 'prop-types'<% if (!includeRedux) { %>
+import { SuspenseWithPerf } from 'reactfire'
 import NavbarWithoutAuth from 'containers/Navbar/NavbarWithoutAuth'<% } %>
 import Navbar from 'containers/Navbar'<% if (includeRedux) { %>
 import { Notifications } from 'modules/notification'<% } %>
-import { makeStyles } from '@material-ui/core/styles'<% if (!includeRedux) { %>
-import LoadingSpinner from 'components/LoadingSpinner'<% } %>
+import { makeStyles } from '@material-ui/core/styles'
 import styles from './CoreLayout.styles'
 
 const useStyles = makeStyles(styles)
@@ -14,14 +14,11 @@ function CoreLayout({ children }) {
 
   return (
     <div className={classes.container}><% if (!includeRedux) { %>
-      <Suspense fallback={<NavbarWithoutAuth />}>
+      <SuspenseWithPerf fallback={<NavbarWithoutAuth />} traceId="load-navbar">
         <Navbar />
-      </Suspense><% } %><% if (includeRedux) { %>
-      <Navbar />
-      <div className={classes.children}>{children}</div><% } %><% if (!includeRedux) { %>
-      <Suspense fallback={<LoadingSpinner />}>
-        <div className={classes.children}>{children}</div>
-      </Suspense><% } %><% if (includeRedux) { %>
+      </SuspenseWithPerf><% } %><% if (includeRedux) { %>
+      <Navbar /><% } %>
+      <div className={classes.children}>{children}</div><% if (includeRedux) { %>
       <Notifications /><% } %>
     </div>
   )
